Move and rename Exception package to root
