+++ src/main/java/se/kth/sda/skeleton/posts/PostController.java
@@ -4,8 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import se.kth.sda.skeleton.posts.exeption.*;
-import se.kth.sda.skeleton.posts.PostService;
 
 import javax.validation.Valid;
 import java.util.List;
